Updated the URLs for TARGET
